refactor: TestController 제거
